analysis: Allow execute variant-export when mode=READ_ONLY #TASK-4488
